Gets activity summary by using calendar service

+++ tntconcept-binnacle/src/main/kotlin/com/autentia/tnt/binnacle/core/domain/DailyWorkingTime.kt
@@ -4,5 +4,5 @@ import java.time.LocalDate
 
 data class DailyWorkingTime(
     val date: LocalDate,
-    val workedHours: Float,
+    val workedHours: Double,
 )
